refactor(harness-cli): own multiagent command contracts
