+++ src/loushang/harness/cli/__init__.py
@@ -130,6 +130,21 @@ from loushang.harness.cli.model_listing import (
     ModelListingResult,
     list_model_entries,
 )
+from loushang.harness.cli.multiagent import (
+    MultiAgentCliCommand,
+    MultiAgentCliUsageError,
+    MultiAgentOutputFormat,
+    MultiAgentRecipesCommand,
+    MultiAgentRunCommand,
+    extract_multiagent_argv,
+    multiagent_recipes_help,
+    multiagent_run_help,
+    parse_multiagent_command,
+    resolve_multiagent_prompt,
+    resolve_multiagent_replicas,
+    write_multiagent_recipe_catalog,
+    write_multiagent_recipe_result,
+)
 from loushang.harness.cli.package_lifecycle import (
     PackageLifecycleError,
     PackageLifecycleRequest,
@@ -317,6 +332,19 @@ __all__ = [
     "ModelListingRequest",
     "ModelListingResult",
     "list_model_entries",
+    "MultiAgentCliCommand",
+    "MultiAgentCliUsageError",
+    "MultiAgentOutputFormat",
+    "MultiAgentRecipesCommand",
+    "MultiAgentRunCommand",
+    "extract_multiagent_argv",
+    "multiagent_recipes_help",
+    "multiagent_run_help",
+    "parse_multiagent_command",
+    "resolve_multiagent_prompt",
+    "resolve_multiagent_replicas",
+    "write_multiagent_recipe_catalog",
+    "write_multiagent_recipe_result",
     "MethodListingError",
     "MethodListingRequest",
     "MethodListingResult",
